requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-rosetta==0.9.8
idna==3.4
polib==1.2.0
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
